Reworked ACNO-files to work with Frozen lake environment.
Overall, POMCP prefered walking into holes and terminating episodes early, because it judged the MDP would give negative reward.
This is caused by the fact that measuring actions have cost, and POMCP did not correcly understand it could always avoid these by not measuring.
The problem seems to be fixed with the following changes:

	* Changed generate_step (& model_step) to have the option to not include measurement costs
		(This makes random rollouts reflect actual value more accurately)
	* Changed ucb_action function (as used by POMCP) to have an option to prefer actions over a certain value (i.e. non-measuring actions) if all else is equal.
		(This seems to help traversal method, thouht I'm not completely sure why it is necessary...)

It might be possible to solve this problem a bit more cleanly...
Status: Running, but does not work well yet, and to be tested on bigger envs.
Current Problems: Does not behave well in big environments (chain 20)